Start removing tight coupling
